requirements.txt
flask
openai
langchain
python-dotenv